refactor: extract command registration from main.ts (#1070)
Pure code-motion refactor. Moves the ~27 `addCommand({...})` registrations
out of `ObsidianGemini.registerUIAndCommands()` into a new
`src/commands/register-commands.ts` exporting `registerCommands(plugin)`.
`main.ts` keeps the ribbon icon, view registration, and the editor-menu
context-menu event, and now calls `registerCommands(this)`.

Command IDs, names, and callback behavior are unchanged. `checkInitialized()`
is widened from `private` to `public` so the extracted module can call it;
the new module imports the plugin type via `import type` to avoid a runtime
import cycle. Removes imports from `main.ts` that are now only used by the
extracted commands.

Fixes #1070
