+++ src/main.ts
@@ -1,6 +1,5 @@
-import { Plugin, WorkspaceLeaf, Editor, MarkdownView, MarkdownFileInfo, Platform } from 'obsidian';
+import { Plugin, WorkspaceLeaf, Platform } from 'obsidian';
 import ObsidianGeminiSettingTab from './ui/settings';
-import { refreshGeminiModelList } from './ui/settings-general';
 import { AgentView, VIEW_TYPE_AGENT } from './ui/agent-view/agent-view';
 import { GeminiDiffView, VIEW_TYPE_DIFF } from './ui/agent-view/gemini-diff-view';
 import { GeminiSummary } from './summary';
@@ -14,7 +13,7 @@ import { ModelManager } from './services/model-manager';
 import { PromptManager, GeminiPrompts } from './prompts';
 import { SelectionRewriter } from './rewrite-selection';
 import { RewriteInstructionsModal } from './ui/rewrite-modal';
-import { UpdateNotificationModal } from './ui/update-notification-modal';
+import { registerCommands } from './commands/register-commands';
 import { SessionManager } from './agent/session-manager';
 import { ToolRegistry } from './tools/tool-registry';
 import { ToolExecutionEngine } from './tools/execution-engine';
@@ -43,7 +42,7 @@ import { BackgroundTaskManager } from './services/background-task-manager';
 import { BackgroundStatusBar } from './services/background-status-bar';
 import { ScheduledTaskManager } from './services/scheduled-task-manager';
 import { HookManager } from './services/hook-manager';
-import { getErrorMessage, getRawErrorMessage } from './utils/error-utils';
+import { getRawErrorMessage } from './utils/error-utils';
 import { t } from './i18n';
 
 export interface RagIndexingSettings {
@@ -291,7 +290,7 @@ export default class ObsidianGemini extends Plugin {
 	 * Check if the plugin is initialized and show a notice if not
 	 * @returns true if initialized, false otherwise
 	 */
-	private checkInitialized(): boolean {
+	public checkInitialized(): boolean {
 		if (!this.isGeminiInitialized) {
 			new Notice(this.getApiKeyErrorMessage());
 			return false;
@@ -340,268 +339,8 @@ export default class ObsidianGemini extends Plugin {
 		this.registerView(VIEW_TYPE_AGENT, (leaf) => (this.agentView = new AgentView(leaf, this)));
 		this.registerView(VIEW_TYPE_DIFF, (leaf) => new GeminiDiffView(leaf, this));
 
-		// Add command
-		this.addCommand({
-			id: 'gemini-scribe-open-agent-view',
-			name: t('command.openAgentView'),
-			callback: () => {
-				if (!this.checkInitialized()) return;
-				this.activateAgentView();
-			},
-		});
-
-		// Refresh remote Gemini model list (bypass 24h cache)
-		this.addCommand({
-			id: 'gemini-scribe-refresh-model-list',
-			name: t('command.refreshModelList'),
-			callback: async () => {
-				if (!this.checkInitialized()) return;
-				await refreshGeminiModelList(this);
-			},
-		});
-
-		// View background tasks
-		this.addCommand({
-			id: 'gemini-scribe-view-background-tasks',
-			name: t('command.viewBackgroundTasks'),
-			callback: async () => {
-				const { BackgroundTasksModal } = await import('./ui/background-tasks-modal');
-				// Command name is unambiguous, so always land on the Tasks tab.
-				// The status-bar entry uses its own context-aware default.
-				new BackgroundTasksModal(this.app, this, 'tasks').open();
-			},
-		});
-
-		// Open scheduler management modal
-		this.addCommand({
-			id: 'gemini-scribe-open-scheduler',
-			name: t('command.openScheduler'),
-			callback: async () => {
-				const { SchedulerManagementModal } = await import('./ui/scheduler-management-modal');
-				new SchedulerManagementModal(this.app, this, 'list').open();
-			},
-		});
-
-		// New scheduled task — jump straight to create form
-		this.addCommand({
-			id: 'gemini-scribe-new-scheduled-task',
-			name: t('command.newScheduledTask'),
-			callback: async () => {
-				const { SchedulerManagementModal } = await import('./ui/scheduler-management-modal');
-				new SchedulerManagementModal(this.app, this, 'create').open();
-			},
-		});
-
-		// Open lifecycle hook management modal
-		this.addCommand({
-			id: 'gemini-scribe-open-hook-manager',
-			name: t('command.openHookManager'),
-			callback: async () => {
-				const { HookManagementModal } = await import('./ui/hook-management-modal');
-				new HookManagementModal(this.app, this, 'list').open();
-			},
-		});
-
-		// New lifecycle hook — jump straight to create form
-		this.addCommand({
-			id: 'gemini-scribe-new-hook',
-			name: t('command.newHook'),
-			callback: async () => {
-				const { HookManagementModal } = await import('./ui/hook-management-modal');
-				new HookManagementModal(this.app, this, 'create').open();
-			},
-		});
-
-		// View scheduled tasks (read-only legacy — kept for backwards compatibility)
-		this.addCommand({
-			id: 'gemini-scribe-view-scheduled-tasks',
-			name: t('command.viewScheduledTasks'),
-			callback: async () => {
-				const { ScheduledTasksModal } = await import('./ui/scheduled-tasks-modal');
-				new ScheduledTasksModal(this.app, this).open();
-			},
-		});
-
-		// Switch project for the current agent session
-		this.addCommand({
-			id: 'gemini-scribe-switch-project',
-			name: t('command.switchProject'),
-			callback: () => {
-				if (!this.checkInitialized()) return;
-				this.activateAgentView();
-				// The agent view's switchProject is triggered via the project badge in the header
-				// or users can click the project indicator once the view is open
-			},
-		});
-
-		// Create a new project
-		this.addCommand({
-			id: 'gemini-scribe-create-project',
-			name: t('command.createProject'),
-			callback: async () => {
-				if (!this.checkInitialized()) return;
-				const folder = this.app.workspace.getActiveFile()?.parent?.path || '';
-				const name = 'New Project';
-				try {
-					const file = await this.projectManager.createProject(folder, name);
-					await this.app.workspace.openLinkText(file.path, '', true);
-					new Notice(t('notice.main.projectCreated', { path: file.path }));
-				} catch (error) {
-					this.logger.error('Failed to create project:', error);
-					new Notice(t('notice.main.projectCreateFailed'));
-				}
-			},
-		});
-
-		// Convert current note to a project
-		this.addCommand({
-			id: 'gemini-scribe-convert-to-project',
-			name: t('command.convertToProject'),
-			editorCallback: async (_editor: Editor, view: MarkdownView | MarkdownFileInfo) => {
-				if (!this.checkInitialized()) return;
-				if (!view.file) return;
-				try {
-					await this.projectManager.convertNoteToProject(view.file);
-					new Notice(t('notice.main.convertedToProject', { name: view.file.basename }));
-				} catch (error) {
-					this.logger.error('Failed to convert note to project:', error);
-					new Notice(t('notice.main.convertToProjectFailed'));
-				}
-			},
-		});
-
-		// Open project settings (the project file itself)
-		this.addCommand({
-			id: 'gemini-scribe-open-project-settings',
-			name: t('command.openProjectSettings'),
-			callback: async () => {
-				if (!this.checkInitialized()) return;
-				const projects = this.projectManager?.discoverProjects() ?? [];
-				if (projects.length === 0) {
-					new Notice(t('notice.main.noProjectsFound'));
-					return;
-				}
-				// If only one project, open it directly
-				if (projects.length === 1) {
-					await this.app.workspace.openLinkText(projects[0].filePath, '', true);
-					return;
-				}
-				// Show picker for multiple projects
-				const { ProjectPickerModal } = await import('./ui/agent-view/project-picker-modal');
-				const modal = new ProjectPickerModal(this.app, this, {
-					onSelect: async (project) => {
-						if (project) {
-							await this.app.workspace.openLinkText(project.filePath, '', true);
-						}
-					},
-				});
-				modal.open();
-			},
-		});
-
-		// Resume the most recent session for a project
-		this.addCommand({
-			id: 'gemini-scribe-resume-project-session',
-			name: t('command.resumeProjectSession'),
-			callback: async () => {
-				if (!this.checkInitialized()) return;
-				const projects = this.projectManager?.discoverProjects() ?? [];
-				if (projects.length === 0) {
-					new Notice(t('notice.main.noProjectsFound'));
-					return;
-				}
-				const { ProjectPickerModal } = await import('./ui/agent-view/project-picker-modal');
-				const modal = new ProjectPickerModal(this.app, this, {
-					onSelect: async (project) => {
-						if (!project) return;
-						// Find most recent session linked to this project
-						const sessions = await this.sessionManager.getRecentAgentSessions(50);
-						const projectSession = sessions.find((s) => s.projectPath === project.filePath);
-						if (projectSession) {
-							await this.activateAgentView();
-							// The agent view will load the session
-							if (this.agentView) {
-								await this.agentView.loadSession(projectSession);
-							}
-						} else {
-							new Notice(t('notice.main.noSessionsForProject', { name: project.name }));
-						}
-					},
-				});
-				modal.open();
-			},
-		});
-
-		// Remove project status from a file
-		this.addCommand({
-			id: 'gemini-scribe-remove-project',
-			name: t('command.removeProject'),
-			editorCallback: async (_editor: Editor, view: MarkdownView | MarkdownFileInfo) => {
-				if (!this.checkInitialized()) return;
-				if (!view.file) return;
-				try {
-					await this.projectManager.removeProject(view.file);
-					new Notice(t('notice.main.projectRemoved', { name: view.file.basename }));
-				} catch (error) {
-					this.logger.error('Failed to remove project:', error);
-					new Notice(t('notice.main.projectRemoveFailed'));
-				}
-			},
-		});
-
-		// Add rewrite command (works with selection or full file)
-		this.addCommand({
-			id: 'gemini-scribe-rewrite-selection',
-			name: t('command.rewriteSelection'),
-			editorCallback: (editor: Editor, _view: MarkdownView | MarkdownFileInfo) => {
-				if (!this.checkInitialized()) return;
-				const selection = editor.getSelection();
-
-				if (!selection || selection.trim().length === 0) {
-					new Notice(t('notice.main.selectTextFirst'));
-					return;
-				}
-
-				const textToRewrite = selection;
-				const isFullFile = false;
-
-				// Show modal for instructions
-				const modal = new RewriteInstructionsModal(
-					this.app,
-					textToRewrite,
-					async (instructions) => {
-						const rewriter = new SelectionRewriter(this);
-						if (isFullFile) {
-							await rewriter.rewriteFullFile(editor, instructions);
-						} else {
-							await rewriter.rewriteSelection(editor, selection, instructions);
-						}
-					},
-					isFullFile
-				);
-				modal.open();
-			},
-		});
-
-		// Add explain selection command
-		this.addCommand({
-			id: 'gemini-scribe-explain-selection',
-			name: t('command.explainSelection'),
-			editorCallback: async (editor: Editor, view: MarkdownView | MarkdownFileInfo) => {
-				if (!this.checkInitialized()) return;
-				await this.selectionActionService.handleExplainSelection(editor, view.file);
-			},
-		});
-
-		// Add ask about selection command
-		this.addCommand({
-			id: 'gemini-scribe-ask-selection',
-			name: t('command.askSelection'),
-			editorCallback: async (editor: Editor, view: MarkdownView | MarkdownFileInfo) => {
-				if (!this.checkInitialized()) return;
-				await this.selectionActionService.handleAskAboutSelection(editor, view.file);
-			},
-		});
+		// Register all command-palette commands (extracted to ./commands/register-commands)
+		registerCommands(this);
 
 		// Add context menu items for selection actions
 		this.registerEvent(
@@ -660,179 +399,6 @@ export default class ObsidianGemini extends Plugin {
 				}
 			})
 		);
-
-		// Add command to view release notes
-		this.addCommand({
-			id: 'gemini-scribe-view-release-notes',
-			name: t('command.viewReleaseNotes'),
-			callback: () => {
-				const modal = new UpdateNotificationModal(this.app, this.manifest.version);
-				modal.open();
-			},
-		});
-
-		// Image generation command (Gemini-only). Registered unconditionally so
-		// the palette entry stays stable across runtime provider switches; the
-		// callback gates on provider before touching the (potentially null)
-		// `imageGeneration` service.
-		this.addCommand({
-			id: 'gemini-scribe-generate-image',
-			name: t('command.generateImage'),
-			callback: async () => {
-				if (!this.checkInitialized()) return;
-				if (this.settings.provider === 'ollama') {
-					new Notice(t('notice.main.imageGenOllama'));
-					return;
-				}
-				if (!this.imageGeneration) {
-					new Notice(t('notice.main.imageGenUnavailable'));
-					return;
-				}
-				const prompt = await this.imageGeneration.promptForImageDescription();
-				if (prompt) {
-					await this.imageGeneration.generateAndInsertImage(prompt);
-				}
-			},
-		});
-
-		// RAG indexing commands. Same pattern as image generation: register
-		// unconditionally and gate at execution time so the palette stays
-		// consistent and the user gets a clear "not available" notice on the
-		// Ollama path (RAG depends on Gemini's File Search Store in Phase 1).
-		this.addCommand({
-			id: 'gemini-scribe-rag-pause',
-			name: t('command.ragPause'),
-			callback: () => {
-				if (this.settings.provider === 'ollama') {
-					new Notice(t('notice.main.ragOllamaUnavailable'));
-					return;
-				}
-				if (!this.ragIndexing) {
-					new Notice(t('notice.main.ragNotEnabled'));
-					return;
-				}
-				if (this.ragIndexing.isPaused()) {
-					new Notice(t('notice.main.ragAlreadyPaused'));
-					return;
-				}
-				if (this.ragIndexing.isIndexing()) {
-					new Notice(t('notice.main.ragCannotPauseWhileIndexing'));
-					return;
-				}
-				this.ragIndexing.pause();
-				new Notice(t('notice.main.ragPaused'));
-			},
-		});
-
-		this.addCommand({
-			id: 'gemini-scribe-rag-resume',
-			name: t('command.ragResume'),
-			callback: () => {
-				if (this.settings.provider === 'ollama') {
-					new Notice(t('notice.main.ragOllamaUnavailable'));
-					return;
-				}
-				if (!this.ragIndexing) {
-					new Notice(t('notice.main.ragNotEnabled'));
-					return;
-				}
-				if (!this.ragIndexing.isPaused()) {
-					new Notice(t('notice.main.ragNotPaused'));
-					return;
-				}
-				this.ragIndexing.resume();
-				new Notice(t('notice.main.ragResumed'));
-			},
-		});
-
-		this.addCommand({
-			id: 'gemini-scribe-rag-status',
-			name: t('command.ragStatus'),
-			callback: async () => {
-				if (this.settings.provider === 'ollama') {
-					new Notice(t('notice.main.ragOllamaUnavailable'));
-					return;
-				}
-				if (!this.ragIndexing) {
-					new Notice(t('notice.main.ragNotEnabled'));
-					return;
-				}
-				// Trigger the same modal as clicking the status bar
-				try {
-					const { openRagStatusModal } = await import('./services/rag-status-bar');
-					await openRagStatusModal(this.app, this.ragIndexing, this.manifest.id);
-				} catch (error) {
-					this.logger.error('RAG Indexing: Failed to open status UI', error);
-					new Notice(t('notice.rag.uiError', { error: getErrorMessage(error) }));
-				}
-			},
-		});
-
-		// Agent session management commands
-		this.addCommand({
-			id: 'gemini-scribe-new-session',
-			name: t('command.newSession'),
-			callback: async () => {
-				if (!this.checkInitialized()) return;
-				// Check if the agent view already exists before activating it.
-				// AgentView.onOpen() automatically creates a default session, so we only
-				// call createNewSession() if the view was already open (user is asking for
-				// a fresh session, not the existing default).
-				const viewAlreadyExists = !!this.agentView;
-				await this.activateAgentView();
-				if (viewAlreadyExists && this.agentView) {
-					await this.agentView.createNewSession();
-				}
-			},
-		});
-
-		this.addCommand({
-			id: 'gemini-scribe-browse-sessions',
-			name: t('command.browseSessions'),
-			callback: async () => {
-				if (!this.checkInitialized()) return;
-				await this.activateAgentView();
-				if (this.agentView) {
-					await this.agentView.showSessionList();
-				}
-			},
-		});
-
-		this.addCommand({
-			id: 'gemini-scribe-link-project',
-			name: t('command.linkProject'),
-			callback: async () => {
-				if (!this.checkInitialized()) return;
-				await this.activateAgentView();
-				if (this.agentView) {
-					this.agentView.switchProject();
-				}
-			},
-		});
-
-		this.addCommand({
-			id: 'gemini-scribe-session-settings',
-			name: t('command.sessionSettings'),
-			callback: async () => {
-				if (!this.checkInitialized()) return;
-				await this.activateAgentView();
-				if (this.agentView) {
-					await this.agentView.showSessionSettings();
-				}
-			},
-		});
-
-		this.addCommand({
-			id: 'gemini-scribe-toggle-plan-mode',
-			name: t('command.togglePlanMode'),
-			callback: async () => {
-				if (!this.checkInitialized()) return;
-				await this.activateAgentView();
-				if (this.agentView) {
-					this.agentView.togglePlanMode();
-				}
-			},
-		});
 	}
 
 	async activateAgentView() {
